fix tests + code quality
