+++ apps/control-plane-backend/alembic/env.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from logging.config import fileConfig
 
+import fred_core.teams.team_metatada_models  # noqa: F401
 from fred_core.sql import make_alembic_env
 from fred_core.users.user_models import UserRow  # noqa: F401
 
@@ -9,7 +10,6 @@ import control_plane_backend.models.agent_instance_models  # noqa: F401
 import control_plane_backend.models.prompt_models  # noqa: F401
 import control_plane_backend.models.purge_queue_models  # noqa: F401
 import control_plane_backend.models.session_metadata_models  # noqa: F401
-import control_plane_backend.models.team_metadata_models  # noqa: F401
 from alembic import context
 from control_plane_backend.config.loader import load_configuration
