+++ apps/control-plane-backend/tests/conftest.py
@@ -16,13 +16,13 @@ def _setup_test_schema() -> None:
     process.  Existing tables are left untouched (checkfirst=True default).
     """
     from fred_core.models.base import Base as FredCoreBase
+    from fred_core.teams import TeamMetadataRow  # noqa: F401
     from fred_core.users.user_models import UserRow  # noqa: F401
 
     import control_plane_backend.models.agent_instance_models  # noqa: F401
     import control_plane_backend.models.prompt_models  # noqa: F401
     import control_plane_backend.models.purge_queue_models  # noqa: F401
     import control_plane_backend.models.session_metadata_models  # noqa: F401
-    import control_plane_backend.models.team_metadata_models  # noqa: F401
     from control_plane_backend.models.base import Base as CPBase
 
     db_path = pathlib.Path("~/.fred/control-plane/control_plane.sqlite3").expanduser()
