+++ libs/fred-core/fred_core/teams/team_metatada_models.py
@@ -1,0 +1,41 @@
+from __future__ import annotations
+
+from datetime import datetime, timezone
+
+from sqlalchemy import BigInteger, Boolean, DateTime, String
+from sqlalchemy.orm import Mapped, mapped_column
+
+from fred_core.models import Base
+
+
+def utcnow() -> datetime:
+    """Return timezone-aware UTC timestamp."""
+    return datetime.now(timezone.utc)
+
+
+class TeamMetadataRow(Base):
+    """ORM model for the ``teammetadata`` table."""
+
+    __tablename__ = "teammetadata"
+
+    id: Mapped[str] = mapped_column(String, primary_key=True)
+    description: Mapped[str | None] = mapped_column(String(180), nullable=True)
+    is_private: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
+    banner_object_storage_key: Mapped[str | None] = mapped_column(
+        String(300), nullable=True
+    )
+    max_resources_storage_size: Mapped[int | None] = mapped_column(
+        BigInteger, nullable=True
+    )
+    current_resources_storage_size: Mapped[int | None] = mapped_column(
+        BigInteger, nullable=False, default=0
+    )
+    created_at: Mapped[datetime] = mapped_column(
+        DateTime(timezone=True), nullable=False, default=utcnow
+    )
+    updated_at: Mapped[datetime] = mapped_column(
+        DateTime(timezone=True),
+        nullable=False,
+        default=utcnow,
+        onupdate=utcnow,
+    )
