+++ libs/fred-core/fred_core/users/user_models.py
@@ -23,5 +23,5 @@ class UserRow(Base):
         DateTime(timezone=True), nullable=True
     )
     current_resources_storage_size: Mapped[int | None] = mapped_column(
-        BigInteger, nullable=True, default=0
+        BigInteger, nullable=False, default=0
     )
